+++ ios/RCTARKitManager.m
@@ -454,8 +454,8 @@ RCT_EXPORT_METHOD(getAnchorPosition:(float)locationLat locationLong:(float)locat
     resolve([[ARKit sharedInstance] getArAnchorPosition:locationLat locationLong:locationLong landmarkLat:landmarkLat landmarkLong:landmarkLong]);
 }
 
-RCT_EXPORT_METHOD(getNewCoords:(float)locationLat locationLong:(float)locationLong distanceKm:(double)distanceKm atBearingDegrees:(double)bearingDegrees resolve:(RCTPromiseResolveBlock)resolve reject:(RCTPromiseRejectBlock)reject) {
-    resolve([[ARKit sharedInstance] coordinateFromCoord:locationLat locationLong:locationLong distanceKm:(double)distanceKm atBearingDegrees:(double)bearingDegrees]);
+RCT_EXPORT_METHOD(getNewCoords:(float)locationLat locationLong:(float)locationLong distanceKm:(float)distanceKm atBearingDegrees:(float)bearingDegrees resolve:(RCTPromiseResolveBlock)resolve reject:(RCTPromiseRejectBlock)reject) {
+    resolve([[ARKit sharedInstance] coordinateFromCoord:locationLat locationLong:locationLong distanceKm:distanceKm atBearingDegrees:bearingDegrees]);
 }
